% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{Function}
\usage{
CountryPopulation(country_name)
}
\arguments{
\item{country}{Country name as a string}
}
\value{
A line graph showing the population estimates
}
\description{
This function generates a plot of the population estimate for a country
}
\examples{
# example code
CountryPopulation('United States of America')
CountryPopulation('Germany')
}
