% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldPopulation.R
\docType{data}
\name{WorldPopulation}
\alias{WorldPopulation}
\title{World Population Estimates}
\format{
A data frame of 16685 observartions
\describe{
   \item{Year}{The year of the population estimate from 1950 - 2020}
   \item{Country}{The country of the poopulation estimate}
   \item{Population}{The estimated population in thousands}
}
}
\source{
United Nations
}
\usage{
WorldPopulation
}
\description{
A collect of population estimates for every country
}
\keyword{datasets}
